requirements.txt
rembg
Pillow
onnxruntime
argparse
numpy<2.0
